+++ packages/apps/job-launcher/server/src/app-init.ts
@@ -1,0 +1,43 @@
+import session from 'express-session';
+import { ConfigService } from '@nestjs/config';
+import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+import { json, urlencoded } from 'body-parser';
+import { useContainer } from 'class-validator';
+import helmet from 'helmet';
+import cookieParser from 'cookie-parser';
+
+import { AppModule } from './app.module';
+import { ConfigNames } from './common/config';
+export default async function init(app: any) {
+  const configService: ConfigService = app.get(ConfigService);
+
+  useContainer(app.select(AppModule), { fallbackOnErrors: true });
+
+  app.use(cookieParser());
+
+  const sessionSecret = configService.get<string>(
+    ConfigNames.SESSION_SECRET,
+    'session-secret',
+  );
+
+  app.use(
+    session({
+      secret: sessionSecret,
+      resave: false,
+      saveUninitialized: false,
+    }),
+  );
+  app.use(json({ limit: '5mb' }));
+  app.use(urlencoded({ limit: '5mb', extended: true }));
+
+  const config = new DocumentBuilder()
+    .addBearerAuth()
+    .setTitle('Job Launcher API')
+    .setDescription('Swagger Job Launcher API')
+    .setVersion('1.0')
+    .build();
+  const document = SwaggerModule.createDocument(app, config);
+  SwaggerModule.setup('swagger', app, document);
+
+  app.use(helmet());
+}
