+++ packages/apps/job-launcher/server/vercel.json
@@ -1,20 +1,16 @@
 {
   "version": 2,
-  "builds": [
-    {
-      "src": "src/main.ts",
-      "use": "@vercel/node"
+  "buildCommand": "yarn workspace @human-protocol/sdk build && yarn build",
+  "outputDirectory": "dist",
+  "functions": {
+    "api/app.ts":{
+      "maxDuration": 300
     }
-  ],
-  "routes": [
+  },
+  "rewrites": [
     {
-      "src": "/(.*)",
-      "dest": "src/main.ts",
-      "headers": {
-        "Access-Control-Allow-Origin": "*",
-        "Access-Control-Allow-Methods": "*",
-        "Access-Control-Allow-Headers": "X-Requested-With,Content-Type,Accept"
-      }
+      "source": "/(.*)",
+      "destination": "/api/app.ts"
     }
   ],
   "crons": [
